refactor: Replace BaseModel with ModifiedBaseModel for consistent alias generation in parsing modules (#66)
